requirements.txt
chromadb
sentence-transformers
python-dotenv
pytz
groq
requests
